python/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-pinecone==0.0.3
pinecone-client==3.0.0
pydantic==2.5.0
httpx==0.26.0
python-dotenv==1.0.0
redis==5.0.0

